Create main.cpp
This is the coding answer for this technical question.
Write a function that takes in a non empty array of distinct integers and an integer representing a target sum. 
If any two numbers in the input array sum up to the target sum, the function should return them in an array. 
If no two numbers sum up to the target sum, the function should return an empty array. 
Assume that there will be **at most one pair of numbers** summing up to the target sum. 

*Testcases:*
Sample input 1: [3,5,-4,8,11,1,-1,6], 10
Sample output 1: [-1,11]

Sample input 2: [2,-7,4,8,-11],56
Sample output 2: []